requirements.txt
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
pillow==10.1.0
PyJWT==2.10.1
drf-yasg==1.21.7
coreapi==2.3.3
markdown==3.5.1
pytz==2023.3
gunicorn==21.2.0
whitenoise==6.6.0 